+++ Source/Atomic/Graphics/Text3D/Text3DFont.h
@@ -66,7 +66,7 @@ public:
     void SetScaledGlyphOffset(const Vector2& offset);
 
     /// Return font face. Pack and render to a texture if not rendered yet. Return null on error.
-    Text3DFontFace* GetFace(int pointSize);
+    Text3DFontFace* GetFace(float pointSize);
 
     /// Return font type.
     Text3DFontType GetFontType() const { return fontType_; }
@@ -81,7 +81,7 @@ public:
     const Vector2& GetScaledGlyphOffset() const { return scaledOffset_; }
 
     /// Return the total effective offset for a point size.
-    IntVector2 GetTotalGlyphOffset(int pointSize) const;
+    IntVector2 GetTotalGlyphOffset(float pointSize) const;
 
     /// Release font faces and recreate them next time when requested. Called when font textures lost or global font properties change.
     void ReleaseFaces();
@@ -90,9 +90,9 @@ private:
     /// Load font glyph offset parameters from an optional XML file. Called internally when loading TrueType fonts.
     void LoadParameters();
     /// Return font face using FreeType. Called internally. Return null on error.
-    Text3DFontFace* GetFaceFreeType(int pointSize);
+    Text3DFontFace* GetFaceFreeType(float pointSize);
     /// Return bitmap font face. Called internally. Return null on error.
-    Text3DFontFace* GetFaceBitmap(int pointSize);
+    Text3DFontFace* GetFaceBitmap(float pointSize);
 
     /// Created faces.
     HashMap<int, SharedPtr<Text3DFontFace> > faces_;
